providers: rename proc_provider to process_provider
