+++ packages/sidesail/lib/config/dependencies.dart
@@ -7,7 +7,7 @@ import 'package:sidesail/pages/tabs/settings/node_settings_tab.dart';
 import 'package:sidesail/providers/balance_provider.dart';
 import 'package:sidesail/providers/bmm_provider.dart';
 import 'package:sidesail/providers/cast_provider.dart';
-import 'package:sidesail/providers/proc_provider.dart';
+import 'package:sidesail/providers/process_provider.dart';
 import 'package:sidesail/providers/transactions_provider.dart';
 import 'package:sidesail/providers/zcash_provider.dart';
 import 'package:sidesail/routing/router.dart';
