+++ packages/sidesail/lib/rpc/rpc.dart
@@ -10,7 +10,7 @@ import 'package:get_it/get_it.dart';
 import 'package:logger/logger.dart';
 import 'package:sidesail/app.dart';
 import 'package:sidesail/pages/tabs/settings/node_settings_tab.dart';
-import 'package:sidesail/providers/proc_provider.dart';
+import 'package:sidesail/providers/process_provider.dart';
 
 // when you implement this class, you should extend a ChangeNotifier, to get
 // a proper implementation of notifyListeners(), e.g:
